Rename audit-log-list to audit-logbook. Inserted a table to test the displaying
