+++ infra/kt_frontend/Tenko/src/app/features/audit/audit.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from "@angular/router";
 
-import { AuditLogListComponent } from "./components/audit-log-list/audit-log-list.component";
+import { AuditLogListComponent } from "./components/audit-logbook/audit-logbook.component";
 
 export const auditRoutes: Routes = [
     { path: 'logs', component: AuditLogListComponent}
