--- infra/kt_frontend/Tenko/src/app/features/audit/components/audit-log-list/audit-log-list.component.html
@@ -1,6 +1,0 @@
-<div>
-    <article *ngFor="let entry of auditLogs$ | async">
-        {{ entry.timestamp }} - {{ entry.audit_id }}
-    </article>
-
-</div>
