+++ infra/kt_frontend/Tenko/src/app/features/audit/components/audit-logbook/audit-logbook.component.html
@@ -1,0 +1,33 @@
+<div *ngIf="(auditLogs$ | async)?.length === 0">No DATA</div>
+
+<ng-container *ngIf="auditLogs$ | async as auditLogs">
+    <table mat-table [dataSource]="auditLogs" class="mat-elevation-z8">
+        <ng-container matColumnDef="timestamp">
+            <th mat-header-cell *matHeaderCellDef>Timestamp</th>
+            <td mat-cell *matCellDef="let entry">{{entry.timestamp}}</td>
+        </ng-container>
+        
+        <ng-container matColumnDef="audit_id">
+            <th mat-header-cell *matHeaderCellDef>Audit ID</th>
+            <td mat-cell *matCellDef="let entry">{{entry.audit_id}}</td>
+        </ng-container>
+        
+        <ng-container matColumnDef="action">
+            <th mat-header-cell *matHeaderCellDef>Action</th>
+            <td mat-cell *matCellDef="let entry">{{entry.action}}</td>
+        </ng-container>
+        
+        <ng-container matColumnDef="user_login">
+            <th mat-header-cell *matHeaderCellDef>User</th>
+            <td mat-cell *matCellDef="let entry">{{entry.user_login}}</td>
+        </ng-container>
+        
+        <ng-container matColumnDef="result">
+            <th mat-header-cell *matHeaderCellDef>Result</th>
+            <td mat-cell *matCellDef="let entry">{{entry.result}}</td>
+        </ng-container>
+        
+        <tr mat-header-row *matHeaderRowDef="displayColumns"></tr>
+        <tr mat-row *matRowDef="let entry; columns: displayColumns;"></tr>
+    </table>
+</ng-container>
