+++ infra/kt_frontend/Tenko/src/app/features/audit/components/audit-logbook/audit-logbook.component.ts
@@ -1,5 +1,6 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { MatTableModule } from '@angular/material/table' 
 
 import { HeaderService } from '../../../../core/services/header.service';
 import { AuditService } from '../../services/audit.service';
@@ -8,19 +9,24 @@ import { AuditEntry } from '../../models/audit.model';
 import { AuditFilter } from '../../models/filter.model';
 
 @Component({
-  selector: 'app-audit-log-list',
+  selector: 'app-audit-logbook',
   imports: [
-    CommonModule
+    CommonModule,
+    MatTableModule
   ],
-  templateUrl: './audit-log-list.component.html',
-  styleUrl: './audit-log-list.component.scss'
+  templateUrl: './audit-logbook.component.html',
+  styleUrl: './audit-logbook.component.scss'
 })
 export class AuditLogListComponent implements OnInit {
   private headerService = inject(HeaderService);
-  private auditService = inject(AuditService)
+  private auditService = inject(AuditService);
 
-  auditLogs$!: Observable<AuditEntry[]>
-  auditFilter?: AuditFilter
+  auditLogs$!: Observable<AuditEntry[]>;
+  auditFilter?: AuditFilter;
+
+  displayColumns: string[] = [
+    'timestamp', 'audit_id', 'action', 'user_login', 'result'
+  ];
 
   ngOnInit(): void {    
     this.headerService.setSubtitle("Journal d'audit")
@@ -30,8 +36,7 @@ export class AuditLogListComponent implements OnInit {
         time_range: {
           start: '2025-05-01 00:00:00',
           end: '2025-05-10 00:00:00'
-        },
-        provider_protocol: ['SAML']
+        }
       }
     }
     this.auditLogs$ = this.auditService.getAuditLogs(this.auditFilter)
